Custom collections : make stored paths relative to home (if they are). Gamelist : don't try to save gamelist when a file without metadata is found ( otherwise it saves everytime ) & trying to secure saving. TextureResources : Avoid crashing if a ImageComponent has negative size ( possible in grid ) SystemView : Realtime position/size calculations of carousel... Not used yet, but I have a project for it....

+++ es-core/src/components/ImageComponent.cpp
@@ -123,7 +123,7 @@ void ImageComponent::resize()
 	mSize[0] = Math::round(mSize.x());
 	mSize[1] = Math::round(mSize.y());
 
-	mTexture->rasterizeAt((size_t)mSize.x(), (size_t)mSize.y());
+	mTexture->rasterizeAt(mSize.x(), mSize.y());
 
 	onSizeChanged();
 }
