+++ es-core/src/resources/TextureResource.h
@@ -56,7 +56,7 @@ public:
 	virtual void initFromMemory(const char* file, size_t length);
 
 	// For scalable source images in textures we want to set the resolution to rasterize at
-	void rasterizeAt(size_t width, size_t height);
+	void rasterizeAt(float width, float height);
 	Vector2f getSourceImageSize() const;
 
 	virtual ~TextureResource();
